Brings back the full 10-level financial architecture (round-off fixes,
financial integrity constraints, accounting engine, receipt allocations,
bank accounts, budgets, credit/debit notes, GST/financial reports, etc.)
that had diverged on origin/claude/production-ready-invoicing-2lSqe,
while preserving features unique to this branch: White/Grey purchase
bill types, the Loans module, and the recent Cash Sales double-counting
fix in the real-time dashboard summary.

Resolved conflicts in: backend/index.js, controller/order.js,
controller/product.js, models/purchaseBill.js, services/dailySummary.js,
validations/order.js, validations/purchaseBill.js, frontend layout.jsx,
orders/create.jsx, purchases/list.jsx, suppliers/list.jsx — keeping both
sides' unique logic and removing accidental duplicate imports/routes
introduced by git's non-conflicting auto-merge.

+++ backend/src/models/orderItems.js
@@ -18,13 +18,13 @@ module.exports = (sequelize, Sequelize) => {
                 allowNull: true
             },
             quantity: {
-                type: Sequelize.DOUBLE
+                type: Sequelize.DECIMAL(15, 2)
             },
             productPrice: {
-                type: Sequelize.DOUBLE
+                type: Sequelize.DECIMAL(15, 2)
             },
             totalPrice: {
-                type: Sequelize.DOUBLE
+                type: Sequelize.DECIMAL(15, 2)
             },
             type: {
                 type: Sequelize.ENUM(Object.values(Enums.product))
